added a config and the message manager
