+++ src/uhc/game/MessageManager.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace uhc\game;
+
+
+class MessageManager
+{
+
+}
